requirements.txt
pyrofork
TgCrypto-pyrofork
pymongo
python-dotenv
psutil
motor
requests
